v0.1.0 - import parser from pydocstyle directly
This is a massive simplification over a bundled copy of the
old pydocstyle parser. I'm sure I tried this back in 2017
when I started this tool, but it didn't work then.

The handling of __all__ was changed in pydocstyle 3.0.0, so
we now require at least that version. This also solves #24.

Bad news: This could be the end of Python 2 support.