Strip out signature from docstring (complex regex)
*Some* patterns of signatures (mostly with respect to variadics) can
be interpreted as incomplete/invalid rST and lead to false positives.

Removing the signature content, without removing the line itself,
should avoid false positives while not requiring adjustments to line
numbering.

Uses a removal regex inspired by that of Sphinx's autodoc. Sphinx's
dates back to 0.6 and has not been modified since so seems rather
stable, but the regex was modified to handle the following:

* handle leading and trailing spaces, Sphinx's (apparently) runs on a
  trimmed string
* ignore spaces specifically, Sphinx matches generalised whitespace
  but the regex *always runs on a single line*, that is not the case
  here so we need a regex which does not eat empty lines and alter
  the line offsets

Background
==========

Habitual use supports the first line of the docstring being an
override to the introspected signature (primarily for C functions
which don't have introspected signatures, but also for situations
where it's not possible or convenient to provide an introspectable
signature).

This use has explicit support in Sphinx's `autodoc`.

Fixes #46